Break discovery tool import cycle
