+++ src/relay_teams/tools/registry/defaults.py
@@ -1,9 +1,12 @@
 # -*- coding: utf-8 -*-
 from __future__ import annotations
 
-from relay_teams.tools.discovery_tools import (
-    DiscoveryToolResolver,
-    TOOLS as DISCOVERY_TOOLS,
+from relay_teams.tools.discovery_tools.resolver import DiscoveryToolResolver
+from relay_teams.tools.discovery_tools.activate_tools import (
+    register as register_activate_tools,
+)
+from relay_teams.tools.discovery_tools.tool_search import (
+    register as register_tool_search,
 )
 from relay_teams.tools.computer_tools import TOOLS as COMPUTER_TOOLS
 from relay_teams.tools.im_tools.im_send import register as register_im_send
@@ -16,6 +19,10 @@ from relay_teams.tools.workspace_tools import TOOLS as WORKSPACE_TOOLS
 
 IM_TOOLS = {
     "im_send": register_im_send,
+}
+DISCOVERY_TOOLS = {
+    "activate_tools": register_activate_tools,
+    "tool_search": register_tool_search,
 }
 HIDDEN_FROM_ROLE_CONFIG: tuple[str, ...] = (
     "im_send",
